Move event record model and store to common
