+++ src/service-worker/debugger-controller.test.ts
@@ -7,8 +7,8 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import {
   newDebuggerAttachedRecord,
   newDebuggerDetachedRecord,
-} from "@/service-worker/event-record.ts";
-import { retrieveAllEventRecords, storeEventRecord } from "@/service-worker/event-store.ts";
+} from "@/common/models/event-record.ts";
+import { retrieveAllEventRecords, storeEventRecord } from "@/common/services/event-store.ts";
 import {
   isDebugging,
   startDebugging,
@@ -16,7 +16,7 @@ import {
   registerDebuggerDetachHandler,
 } from "./debugger-controller.ts";
 
-vi.mock("@/service-worker/event-store.ts", () => ({
+vi.mock("@/common/services/event-store.ts", () => ({
   retrieveAllEventRecords: vi.fn(),
   storeEventRecord: vi.fn(),
 }));
